requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.34.0
